refactor: 알고리즘 분석 UI/UX 개선 및 백엔드 리팩토링
## 백엔드 (services/ai-engine/)

### Router 분리 (Issue #7)
- main.py를 1,701줄에서 161줄로 축소
- 기능별 router 파일 7개 생성:
  - routers/strategy.py: 분석 관련 엔드포인트
  - routers/reports.py: 보고서 조회/저장
  - routers/market.py: 시장/기업 데이터
  - routers/macro.py: 매크로 데이터
  - routers/screener.py: 스크리너
  - routers/portfolio.py: 포트폴리오
  - routers/alpha_hr.py: 투자자 매매동향
- models/requests.py: Pydantic 모델 통합

### 거래량/체결강도 분석 추가
- indicators.py: analyze_recent_volatility() 개선
  - 거래량 수준 (높음/보통/낮음)
  - 거래량 추이 (증가/감소/보합)
  - 평균 거래대금
  - 거래대금 추이
  - 체결강도 추정치 (상승일/하락일 거래량 기준)
- chart_analyst.py: 보고서에 거래 현황 테이블 추가

### 종합분석 통합
- "심층분석" → "종합분석"으로 명칭 변경
- 종합분석 = 알고리즘 분석 요약 + LLM 심층 분석 (하나의 보고서)

## 프론트엔드

### 알고리즘 보고서 UI 개선 (algo-report-view.tsx)
- 화려한 대시보드 스타일 → 전문 보고서 문서 스타일
- 무채색 기반, 여백/타이포그래피 중심 디자인
- 숫자 가독성 개선 (오른쪽 정렬, 볼드 처리)
- 타임프레임별 전략:
  - 추세/전략 강조 (회색 배경 박스)
  - 현재가를 테이블에 추가 (진입/손절/목표 비교 가능)
  - 지지선/저항선 깔끔한 표시
- 거래 현황 요약 표 추가:
  - 5일 누적 변동 / 평균 일간 변동
  - 거래량 수준 / 거래량 추이
  - 평균 거래대금 / 거래대금 추이
  - 체결강도 추정

### 파싱 로직 개선 (parse-algo-report.ts)
- 현재가 파싱 정규식 수정 (blockquote 지원)
- 거래량/체결강도 데이터 추출
- 지지선/저항선/목표가 파싱 개선

### 의존성
- remark-gfm 설치: 마크다운 테이블 렌더링 지원

## 문서

- docs/MULTI_TIMEFRAME_STRATEGY.md 추가
  - 멀티 타임프레임 전략 상세 설명
  - 목표가 계산 근거 (5가지 저항선 탐지 방법)
  - 투자 목적별 전략 가이드
  - 실전 활용 예시
