+++ components/features/reports/algo-report-view.tsx
@@ -2,127 +2,96 @@
 
 import React from 'react';
 import ReactMarkdown from 'react-markdown';
+import remarkGfm from 'remark-gfm';
 import { parseAlgoReport, ParsedAlgoReport, ParsedTimeframe, ParsedIndicator } from '@/lib/utils/parse-algo-report';
 
-const TIMEFRAME_COLORS: Record<string, { border: string; bg: string; text: string }> = {
-    '주봉': { border: 'border-l-blue-500', bg: 'bg-blue-50', text: 'text-blue-700' },
-    '일봉': { border: 'border-l-emerald-500', bg: 'bg-emerald-50', text: 'text-emerald-700' },
-    '15분봉': { border: 'border-l-amber-500', bg: 'bg-amber-50', text: 'text-amber-700' },
-    '5분봉': { border: 'border-l-purple-500', bg: 'bg-purple-50', text: 'text-purple-700' },
-};
-
-function StatusBadge({ status, children }: { status: 'bullish' | 'bearish' | 'neutral'; children: React.ReactNode }) {
-    const colors = {
-        bullish: 'bg-emerald-100 text-emerald-800 border-emerald-200',
-        bearish: 'bg-red-100 text-red-800 border-red-200',
-        neutral: 'bg-gray-100 text-gray-700 border-gray-200',
-    };
+function TimeframeCard({ tf, currentPrice }: { tf: ParsedTimeframe; currentPrice?: string }) {
     return (
-        <span className={`inline-flex items-center px-2.5 py-1 rounded-full text-xs font-medium border ${colors[status]}`}>
-            {children}
-        </span>
-    );
-}
-
-function PriceBox({ label, value, variant }: { label: string; value: string; variant: 'entry' | 'stop' | 'target' }) {
-    const styles = {
-        entry: 'bg-blue-50 border-blue-200 text-blue-900',
-        stop: 'bg-red-50 border-red-200 text-red-900',
-        target: 'bg-emerald-50 border-emerald-200 text-emerald-900',
-    };
-    const icons = { entry: '📥', stop: '⛔', target: '🎯' };
-    return (
-        <div className={`flex flex-col items-center p-3 rounded-lg border ${styles[variant]}`}>
-            <span className="text-xs text-gray-500 mb-1">{icons[variant]} {label}</span>
-            <span className="font-bold text-sm">{value}</span>
-        </div>
-    );
-}
-
-function IndicatorChip({ indicator }: { indicator: ParsedIndicator }) {
-    const colors = {
-        bullish: 'bg-emerald-50 text-emerald-700',
-        bearish: 'bg-red-50 text-red-700',
-        neutral: 'bg-gray-50 text-gray-600',
-    };
-    return (
-        <div className={`inline-flex items-center gap-1.5 px-3 py-1.5 rounded-md text-sm ${colors[indicator.status]}`}>
-            <span className="font-medium text-gray-500">{indicator.name}</span>
-            <span className="font-bold">{indicator.value}</span>
-            {indicator.icon && <span>{indicator.icon}</span>}
-        </div>
-    );
-}
-
-function VolatilityStrip({ days }: { days: ParsedAlgoReport['volatility']['days'] }) {
-    const statusColors = {
-        surge: 'bg-emerald-500 text-white',
-        up: 'bg-emerald-100 text-emerald-800',
-        flat: 'bg-gray-100 text-gray-600',
-        down: 'bg-red-100 text-red-800',
-        crash: 'bg-red-500 text-white',
-    };
-    return (
-        <div className="flex gap-2">
-            {days.map((day, i) => (
-                <div key={i} className={`flex-1 p-2.5 rounded-lg text-center ${statusColors[day.status]}`}>
-                    <div className="text-[10px] opacity-70">{day.date.slice(5)}</div>
-                    <div className="text-sm font-bold">{day.changePct > 0 ? '+' : ''}{day.changePct.toFixed(1)}%</div>
-                    <div className="text-[10px] opacity-70">{day.volumeRatio}</div>
-                </div>
-            ))}
-        </div>
-    );
-}
-
-function TimeframeCard({ tf }: { tf: ParsedTimeframe }) {
-    const colors = TIMEFRAME_COLORS[tf.name] || TIMEFRAME_COLORS['일봉'];
-
-    return (
-        <div className={`border-l-4 ${colors.border} bg-white rounded-lg shadow-sm border border-gray-200 overflow-hidden`}>
-            {/* Header */}
-            <div className={`px-4 py-3 ${colors.bg} flex items-center justify-between`}>
-                <div className="flex items-center gap-2">
-                    <span className="text-lg">{tf.icon}</span>
-                    <div>
-                        <h4 className={`font-bold text-sm ${colors.text}`}>{tf.name}</h4>
-                        <p className="text-xs text-gray-500">{tf.subtitle}</p>
+        <div className="bg-white border-b border-gray-300 pb-6 mb-6">
+            {/* Header - Simple and Clean */}
+            <div className="mb-4">
+                <h4 className="text-base font-bold text-gray-900 mb-1">
+                    {tf.name} <span className="text-sm font-normal text-gray-500">({tf.subtitle})</span>
+                </h4>
+                <div className="bg-gray-50 border border-gray-300 px-3 py-2 rounded">
+                    <div className="text-sm">
+                        <span className="font-semibold text-gray-700">추세:</span> <span className="font-bold text-gray-900">{tf.trend}</span>
+                        <span className="mx-2 text-gray-400">|</span>
+                        <span className="font-semibold text-gray-700">전략:</span> <span className="font-bold text-gray-900">{tf.strategy}</span>
                     </div>
                 </div>
-                <StatusBadge status={tf.trendStatus}>{tf.trend}</StatusBadge>
             </div>
 
-            {/* Indicators */}
-            <div className="px-4 py-3">
-                {tf.indicators.length > 0 ? (
-                    <div className="flex flex-wrap gap-2 mb-3">
+            {/* Trading Table */}
+            <table className="w-full text-sm mb-4 border-collapse">
+                <thead>
+                    <tr className="border-b-2 border-gray-400">
+                        <th className="text-left py-2 px-3 font-semibold text-gray-700">구분</th>
+                        <th className="text-right py-2 px-3 font-semibold text-gray-700">가격</th>
+                    </tr>
+                </thead>
+                <tbody className="divide-y divide-gray-200">
+                    {currentPrice && (
+                        <tr className="bg-blue-50">
+                            <td className="py-2 px-3 font-semibold text-gray-700">현재가</td>
+                            <td className="py-2 px-3 text-right font-bold text-lg text-blue-900">{currentPrice}</td>
+                        </tr>
+                    )}
+                    {tf.entry && (
+                        <tr>
+                            <td className="py-2 px-3 text-gray-600">진입</td>
+                            <td className="py-2 px-3 text-right font-bold text-blue-700">{tf.entry}</td>
+                        </tr>
+                    )}
+                    {tf.stopLoss && (
+                        <tr>
+                            <td className="py-2 px-3 text-gray-600">손절</td>
+                            <td className="py-2 px-3 text-right font-bold text-red-600">{tf.stopLoss}</td>
+                        </tr>
+                    )}
+                    {(tf.targets && tf.targets.length > 0) ? (
+                        <tr>
+                            <td className="py-2 px-3 text-gray-600">목표 (1차 → 2차 → 3차)</td>
+                            <td className="py-2 px-3 text-right font-bold text-emerald-700">{tf.targets.join(' → ')}</td>
+                        </tr>
+                    ) : tf.target ? (
+                        <tr>
+                            <td className="py-2 px-3 text-gray-600">목표</td>
+                            <td className="py-2 px-3 text-right font-bold text-emerald-700">{tf.target}</td>
+                        </tr>
+                    ) : null}
+                </tbody>
+            </table>
+
+            {/* Support/Resistance */}
+            {(tf.supports || tf.resistances) && (
+                <div className="space-y-2 text-sm">
+                    {tf.supports && (
+                        <div className="flex">
+                            <span className="w-20 text-gray-600 flex-shrink-0">지지선:</span>
+                            <span className="text-gray-800">{tf.supports}</span>
+                        </div>
+                    )}
+                    {tf.resistances && (
+                        <div className="flex">
+                            <span className="w-20 text-gray-600 flex-shrink-0">저항선:</span>
+                            <span className="text-gray-800">{tf.resistances}</span>
+                        </div>
+                    )}
+                </div>
+            )}
+
+            {/* Indicators - Simple List */}
+            {tf.indicators.length > 0 && (
+                <div className="mt-3 pt-3 border-t border-gray-200">
+                    <div className="text-xs text-gray-500 mb-1">기술적 지표</div>
+                    <div className="flex flex-wrap gap-x-4 gap-y-1 text-sm text-gray-700">
                         {tf.indicators.map((ind, i) => (
-                            <IndicatorChip key={i} indicator={ind} />
+                            <span key={i}>{ind.name}: <strong>{ind.value}</strong></span>
                         ))}
                     </div>
-                ) : null}
-
-                {/* Strategy */}
-                <div className="flex items-center gap-2 mb-3">
-                    <span className="text-xs font-medium text-gray-500 uppercase tracking-wider">Strategy</span>
-                    <StatusBadge status={tf.strategyStatus}>{tf.strategy}</StatusBadge>
                 </div>
-
-                {/* Prices */}
-                {(tf.entry || tf.stopLoss || tf.target) && (
-                    <div className="grid grid-cols-3 gap-2">
-                        {tf.entry && <PriceBox label="진입" value={tf.entry} variant="entry" />}
-                        {tf.stopLoss && <PriceBox label="손절" value={tf.stopLoss} variant="stop" />}
-                        {tf.target && <PriceBox label="목표" value={tf.target} variant="target" />}
-                    </div>
-                )}
-
-                {tf.extra && (
-                    <div className="mt-3 p-3 bg-amber-50 border border-amber-200 rounded-lg text-sm text-amber-800">
-                        {tf.extra}
-                    </div>
-                )}
-            </div>
+            )}
         </div>
     );
 }
@@ -138,119 +107,172 @@ export function AlgoReportView({ analysis }: AlgoReportViewProps) {
     if (!parsed.timeframes.length && !parsed.volatility.days.length) {
         return (
             <div className="prose prose-gray max-w-none">
-                <ReactMarkdown>{analysis}</ReactMarkdown>
+                <ReactMarkdown remarkPlugins={[remarkGfm]}>{analysis}</ReactMarkdown>
             </div>
         );
     }
 
     return (
-        <div className="space-y-6">
-            {/* Header */}
-            <div className="flex items-center justify-between">
-                <h2 className="text-xl font-bold text-gray-900">
-                    ⚡ 기술적 분석
+        <div className="bg-white">
+            {/* Header - Document Style */}
+            <div className="border-b-2 border-gray-900 pb-3 mb-6">
+                <h2 className="text-xl font-bold text-gray-900 mb-1">
+                    기술적 분석 보고서
                 </h2>
                 {parsed.currentPrice && (
-                    <span className="text-lg font-bold text-gray-900">
-                        현재가 {parsed.currentPrice}
-                    </span>
+                    <div className="text-sm text-gray-600">
+                        현재가: <span className="font-bold text-gray-900">{parsed.currentPrice}</span>
+                    </div>
                 )}
             </div>
 
-            {/* Stress Alert */}
+            {/* Stress Alert - Simple Table */}
             {parsed.stressAlert && (
-                <div className={`p-4 rounded-lg border ${
-                    parsed.stressAlert.level === 'EXTREME' ? 'bg-red-50 border-red-300' :
-                    parsed.stressAlert.level === 'HIGH' ? 'bg-orange-50 border-orange-300' :
-                    'bg-yellow-50 border-yellow-300'
-                }`}>
-                    <div className="flex items-center gap-2 font-bold text-sm mb-2">
-                        {parsed.stressAlert.icon} 시장 스트레스: {parsed.stressAlert.level} ({parsed.stressAlert.score}/100)
+                <div className="mb-6">
+                    <div className="text-sm font-semibold text-gray-700 mb-2">
+                        시장 상태: {parsed.stressAlert.level} (스트레스 지수 {parsed.stressAlert.score}/100)
                     </div>
-                    {parsed.stressAlert.reasons.map((r, i) => (
-                        <div key={i} className="text-sm text-gray-700">- {r}</div>
-                    ))}
+                    <ul className="list-disc list-inside text-sm text-gray-700 space-y-1">
+                        {parsed.stressAlert.reasons.map((r, i) => (
+                            <li key={i}>{r}</li>
+                        ))}
+                    </ul>
+                    <hr className="mt-4 border-gray-300" />
                 </div>
             )}
 
-            {/* 5-Day Volatility */}
-            {parsed.volatility.days.length > 0 && (
-                <div>
-                    <h3 className="text-sm font-semibold text-gray-500 uppercase tracking-wider mb-3">
-                        최근 5일 변동
+            {/* 5-Day Volatility Table */}
+            {parsed.volatility.rawMarkdown ? (
+                <div className="mb-6">
+                    <h3 className="text-lg font-bold text-gray-900 mb-3 border-b border-gray-400 pb-2">
+                        최근 5일 거래 현황
                     </h3>
-                    <VolatilityStrip days={parsed.volatility.days} />
-                    <div className="flex gap-4 mt-2 text-xs text-gray-500">
-                        {parsed.volatility.cumulativeChange && (
-                            <span>누적: <b className="text-gray-700">{parsed.volatility.cumulativeChange}</b></span>
-                        )}
-                        {parsed.volatility.avgDaily && (
-                            <span>평균: <b className="text-gray-700">{parsed.volatility.avgDaily}</b></span>
-                        )}
-                        {parsed.volatility.level && (
-                            <StatusBadge status={
-                                parsed.volatility.level === 'HIGH' ? 'bearish' :
-                                parsed.volatility.level === 'LOW' ? 'bullish' : 'neutral'
-                            }>
-                                {parsed.volatility.level}
-                            </StatusBadge>
-                        )}
-                    </div>
-                </div>
-            )}
 
-            {/* Smart Summary */}
+                    {/* Daily Table */}
+                    <div className="overflow-x-auto mb-4">
+                        <ReactMarkdown
+                            remarkPlugins={[remarkGfm]}
+                            components={{
+                                h2: ({node, ...props}) => <div />,
+                                table: ({node, ...props}) => (
+                                    <table className="w-full border-collapse text-sm" {...props} />
+                                ),
+                                thead: ({node, ...props}) => <thead {...props} />,
+                                tbody: ({node, ...props}) => (
+                                    <tbody className="divide-y divide-gray-200" {...props} />
+                                ),
+                                th: ({node, ...props}) => (
+                                    <th className="px-3 py-2 text-left font-semibold text-gray-700 border-b-2 border-gray-400" {...props} />
+                                ),
+                                td: ({node, ...props}) => (
+                                    <td className="px-3 py-2 text-gray-800" {...props} />
+                                ),
+                                p: ({node, ...props}) => <div />, // Skip p tags
+                                strong: ({node, ...props}) => <div />, // Skip summary lines
+                            }}
+                        >
+                            {parsed.volatility.rawMarkdown.split('\n').filter(line =>
+                                line.includes('|') || line.includes('---')
+                            ).join('\n')}
+                        </ReactMarkdown>
+                    </div>
+
+                    {/* Summary Table */}
+                    <table className="w-full text-sm border-collapse mb-3">
+                        <tbody className="divide-y divide-gray-200">
+                            <tr>
+                                <td className="py-2 px-3 text-gray-600 w-1/3">5일 누적 변동</td>
+                                <td className="py-2 px-3 font-bold text-gray-900">{parsed.volatility.cumulativeChange}</td>
+                                <td className="py-2 px-3 text-gray-600 w-1/3">평균 일간 변동</td>
+                                <td className="py-2 px-3 font-bold text-gray-900">{parsed.volatility.avgDaily}</td>
+                            </tr>
+                            {(parsed.volatility.volumeLevel || parsed.volatility.volumeTrend) && (
+                                <tr>
+                                    <td className="py-2 px-3 text-gray-600">거래량 수준</td>
+                                    <td className="py-2 px-3 font-bold text-gray-900">{parsed.volatility.volumeLevel}</td>
+                                    <td className="py-2 px-3 text-gray-600">거래량 추이</td>
+                                    <td className="py-2 px-3 font-bold text-gray-900">{parsed.volatility.volumeTrend}</td>
+                                </tr>
+                            )}
+                            {(parsed.volatility.avgTradeAmount || parsed.volatility.tradeAmountTrend) && (
+                                <tr>
+                                    <td className="py-2 px-3 text-gray-600">평균 거래대금</td>
+                                    <td className="py-2 px-3 font-bold text-gray-900">{parsed.volatility.avgTradeAmount}</td>
+                                    <td className="py-2 px-3 text-gray-600">거래대금 추이</td>
+                                    <td className="py-2 px-3 font-bold text-gray-900">{parsed.volatility.tradeAmountTrend}</td>
+                                </tr>
+                            )}
+                            {parsed.volatility.buyingPressure && (
+                                <tr>
+                                    <td className="py-2 px-3 text-gray-600">체결강도 추정</td>
+                                    <td className="py-2 px-3 font-bold text-gray-900" colSpan={3}>
+                                        {parsed.volatility.buyingPressure} {parsed.volatility.buyingPressureStatus}
+                                        <span className="text-xs text-gray-500 ml-2">※ 상승일/하락일 거래량 기준 추정치</span>
+                                    </td>
+                                </tr>
+                            )}
+                        </tbody>
+                    </table>
+                    <hr className="border-gray-300" />
+                </div>
+            ) : null}
+
+            {/* Smart Summary - Clean */}
             {parsed.smartSummary.recommendation && (
-                <div className="p-4 bg-gray-50 rounded-lg border border-gray-200">
-                    <div className="flex items-center gap-2 mb-2">
-                        <span className="text-2xl">{parsed.smartSummary.icon}</span>
-                        <span className="text-lg font-bold text-gray-900">{parsed.smartSummary.recommendation}</span>
+                <div className="mb-6">
+                    <div className="text-base font-bold text-gray-900 mb-2">
+                        종합 의견: {parsed.smartSummary.recommendation}
                     </div>
                     {parsed.smartSummary.reason && (
-                        <p className="text-sm text-gray-600 mb-2">{parsed.smartSummary.reason}</p>
-                    )}
-                    {parsed.smartSummary.others.length > 0 && (
-                        <div className="flex flex-wrap gap-2 mt-2">
-                            {parsed.smartSummary.others.map((o, i) => (
-                                <span key={i} className="text-xs bg-white border border-gray-200 px-2 py-1 rounded-full text-gray-600">{o}</span>
-                            ))}
-                        </div>
+                        <p className="text-sm text-gray-700 mb-2">{parsed.smartSummary.reason}</p>
                     )}
                     {parsed.smartSummary.cautions.length > 0 && (
-                        <div className="mt-2 text-xs text-amber-700">
-                            ⚠️ {parsed.smartSummary.cautions.join(' / ')}
+                        <div className="text-sm text-gray-600">
+                            주의사항: {parsed.smartSummary.cautions.join(', ')}
                         </div>
                     )}
+                    <hr className="mt-4 border-gray-300" />
                 </div>
             )}
 
-            {/* Timeframe Cards */}
-            <div>
-                <h3 className="text-sm font-semibold text-gray-500 uppercase tracking-wider mb-3">
-                    멀티 타임프레임 분석
+            {/* Timeframe Analysis - Document Style */}
+            <div className="mb-6">
+                <h3 className="text-lg font-bold text-gray-900 mb-4 border-b border-gray-400 pb-2">
+                    타임프레임별 전략
                 </h3>
-                <div className="grid grid-cols-1 lg:grid-cols-2 gap-4">
-                    {parsed.timeframes.map((tf, i) => (
-                        <TimeframeCard key={i} tf={tf} />
-                    ))}
-                </div>
+                {parsed.timeframes.map((tf, i) => (
+                    <TimeframeCard key={i} tf={tf} currentPrice={parsed.currentPrice} />
+                ))}
             </div>
 
-            {/* Allocation */}
+            {/* Allocation - Simple Table */}
             {parsed.allocation.length > 0 && (
                 <div>
-                    <h3 className="text-sm font-semibold text-gray-500 uppercase tracking-wider mb-3">
-                        포지션 배분 가이드
+                    <h3 className="text-lg font-bold text-gray-900 mb-3 border-b border-gray-400 pb-2">
+                        포지션별 배분 가이드
                     </h3>
-                    <div className="grid grid-cols-4 gap-2">
-                        {parsed.allocation.map((a, i) => (
-                            <div key={i} className="text-center p-3 bg-gray-50 rounded-lg border border-gray-200">
-                                <div className="text-sm font-medium text-gray-900">{a.style}</div>
-                                <div className="text-lg font-bold text-gray-900 my-1">{a.weight}</div>
-                                <div className="text-xs text-gray-500">{a.timeframe}</div>
-                                <div className="text-xs text-gray-400">{a.period}</div>
-                            </div>
-                        ))}
+                    <table className="w-full text-sm border-collapse mb-3">
+                        <thead>
+                            <tr className="border-b-2 border-gray-400">
+                                <th className="text-left py-2 px-3 font-semibold text-gray-700">투자 스타일</th>
+                                <th className="text-center py-2 px-3 font-semibold text-gray-700">비중</th>
+                                <th className="text-center py-2 px-3 font-semibold text-gray-700">타임프레임</th>
+                                <th className="text-right py-2 px-3 font-semibold text-gray-700">보유 기간</th>
+                            </tr>
+                        </thead>
+                        <tbody className="divide-y divide-gray-200">
+                            {parsed.allocation.map((a, i) => (
+                                <tr key={i}>
+                                    <td className="py-2 px-3 text-gray-800">{a.style}</td>
+                                    <td className="py-2 px-3 text-center font-bold text-gray-900">{a.weight}</td>
+                                    <td className="py-2 px-3 text-center text-gray-700">{a.timeframe}</td>
+                                    <td className="py-2 px-3 text-right text-gray-600">{a.period}</td>
+                                </tr>
+                            ))}
+                        </tbody>
+                    </table>
+                    <div className="text-sm text-gray-600 italic">
+                        * 리스크 관리: 각 타임프레임의 손절가를 반드시 준수하세요.
                     </div>
                 </div>
             )}
