+++ docs/MULTI_TIMEFRAME_STRATEGY.md
@@ -1,0 +1,177 @@
+# 멀티 타임프레임 전략 설명
+
+## 개요
+
+StockIQ의 알고리즘 분석은 **멀티 타임프레임 전략**을 사용합니다. 같은 종목이라도 투자 기간에 따라 다른 진입가, 손절가, 목표가를 제시합니다.
+
+## 왜 타임프레임마다 목표가가 다른가?
+
+### 1. 사용하는 데이터가 다름
+
+각 타임프레임은 완전히 다른 가격 데이터를 분석합니다:
+
+- **주봉**: 최근 52주 (1년) 데이터 - 주 단위 캔들
+- **일봉**: 최근 60~120일 데이터 - 일 단위 캔들
+- **15분봉**: 최근 며칠 데이터 - 15분 단위 캔들
+- **5분봉**: 최근 1~2일 데이터 - 5분 단위 캔들
+
+### 2. 시간 범위가 다름
+
+| 타임프레임 | 데이터 범위 | 스윙 포인트 기준 |
+|-----------|------------|-----------------|
+| 주봉 | 최근 1년 | 수개월 전 고점/저점 |
+| 일봉 | 최근 2~4개월 | 수주~수개월 전 고점/저점 |
+| 15분봉 | 최근 며칠 | 수시간~하루 전 고점/저점 |
+| 5분봉 | 최근 1~2일 | 수분~수시간 전 고점/저점 |
+
+### 3. 발견되는 저항선이 다름
+
+**예시: 삼성전자 현재가 164,400원**
+
+**주봉에서 보면:**
+```
+최근 1년간 전고점: 169,400원 (2024년 12월)
+52주 이동평균선: 84,040원
+→ 목표: 169,400원 → 180,000원 → 190,000원 (장기 저항선)
+```
+
+**일봉에서 보면:**
+```
+최근 2개월 전고점: 169,400원
+볼린저 밴드 상단: 172,467원
+→ 목표: 169,400원 → 172,467원 → 180,000원 (중기 저항선)
+```
+
+**15분봉에서 보면:**
+```
+오늘 고점: 165,975원
+당일 VWAP: 161,263원
+→ 목표: 165,975원 → 167,992원 (당일 저항선)
+```
+
+**5분봉에서 보면:**
+```
+최근 1시간 고점: 166,216원
+12분봉 이동평균: 163,942원
+→ 목표: 166,216원 (초단기 저항선)
+```
+
+## 목표가 계산 근거
+
+목표가는 `analyzers/levels.py`의 `pick_entry_stop_target()` 함수에서 계산됩니다.
+
+### 저항선 탐지 (5가지 방법)
+
+1. **스윙 하이 (전고점)**
+   - 과거 가격에서 고점들을 찾아 클러스터링
+   - 여러 번 반복된 가격대 = 강한 저항선
+
+2. **이동평균선**
+   - MA13, MA26, MA52 등
+   - 현재가보다 높으면 저항선
+
+3. **볼린저 밴드 상단**
+   - 과열 구간 표시
+
+4. **심리적 가격**
+   - 10,000원, 50,000원, 100,000원 등 큰 단위
+   - 투자자들이 매도 주문 많이 거는 가격
+
+5. **피보나치 되돌림**
+   - 최근 고점-저점 사이 비율 (23.6%, 38.2%, 50%, 61.8%)
+
+### 목표가 선택 로직
+
+```python
+# 현재가보다 0.5% 이상 높은 저항선들을 가까운 순서대로 선택
+targets = []
+for r in resistances:
+    if r["price"] > current_price * 1.005:
+        targets.append(r["price"])
+    if len(targets) >= 3:
+        break
+
+# 저항선이 부족하면 5% 간격으로 자동 생성
+while len(targets) < 3:
+    last = targets[-1] if targets else current_price
+    targets.append(round(last * 1.05))
+```
+
+## 투자 목적별 전략
+
+| 타임프레임 | 보유 기간 | 목적 | 목표가 특징 |
+|-----------|----------|------|-----------|
+| 주봉 | 6개월~1년 | 장기 포지션 | 큰 상승 기대, 심리적 가격 중심 |
+| 일봉 | 3~10일 | 스윙 매매 | 중기 저항선, 분할 익절 |
+| 15분봉 | 1~3시간 | 데이 트레이딩 | 당일 고점, 빠른 수익 실현 |
+| 5분봉 | 5~15분 | 스캘핑 | 초단기 저항, 소액 빠른 회전 |
+
+## 분할 익절 전략
+
+목표가가 여러 개인 이유:
+
+- **1차 목표 (169,400원)**: 포지션의 30~50% 익절
+- **2차 목표 (172,467원)**: 추가 30% 익절
+- **3차 목표 (180,000원)**: 나머지 전량 익절
+
+**장점:**
+- 리스크 관리 (수익 일부는 확보)
+- 수익 극대화 (상승 추세 지속 시 큰 수익)
+- 심리적 안정 (일부 익절로 불안감 해소)
+
+## 실전 활용 예시
+
+**투자자 프로필별 전략:**
+
+### 장기 투자자 (주봉)
+```
+현재가: 164,400원
+진입: 163,350~164,400원
+손절: 135,706원
+목표: 169,400원 → 180,000원 → 190,000원
+→ 6개월~1년 보유, 큰 상승 기대
+```
+
+### 스윙 트레이더 (일봉)
+```
+현재가: 164,400원
+진입: 163,350~164,400원
+손절: 158,968원
+목표: 169,400원 → 172,467원 → 180,000원
+→ 3~10일 보유, 중기 저항선 익절
+```
+
+### 데이 트레이더 (15분봉)
+```
+현재가: 164,400원
+진입: 163,770~164,400원
+손절: 160,842원
+목표: 165,975원 → 167,992원
+→ 당일 매매, 작은 수익 여러 번
+```
+
+### 스캘퍼 (5분봉)
+```
+현재가: 164,400원
+진입: 대기 (거래량 3배 이상 시)
+손절: 162,692원
+목표: 166,216원
+→ 5~15분, 0.5~1% 수익 빠른 회전
+```
+
+## 주의사항
+
+1. **각 타임프레임은 독립적**: 주봉 목표와 일봉 목표는 서로 다른 분석
+2. **동시에 여러 전략 가능**: 장기 포지션 + 단기 매매 병행
+3. **시장 상황 고려**: 높은 변동성 시 단기 전략 비활성화
+4. **반드시 손절가 준수**: 각 타임프레임의 손절가는 엄격히 지켜야 함
+
+## 기술 구현
+
+- **파일**: `services/ai-engine/analyzers/levels.py`
+- **함수**: `compute_levels()`, `pick_entry_stop_target()`
+- **사용 알고리즘**: 스윙 포인트 탐지, 클러스터링, 피보나치, 이동평균
+
+---
+
+**멀티 타임프레임 전략은 프로 트레이더들이 사용하는 정통 기술적 분석 방법입니다.**
